+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Fixed
 - Displaying sport for athletics events [#3621](https://github.com/rokwire/illinois-app/issues/3621).
+### Added
+- Added published flag in create/update event panel [#3623](https://github.com/rokwire/illinois-app/issues/3623).
 
 ## [5.0.52] - 2023-08-22
 ### Fixed